+++ apps/docs/src/content/04-components/data-visualisation/cartesian-chart/examples/synced.tsx
@@ -1,90 +1,102 @@
 import {
-  Area,
-  CartesianChart,
-  CartesianGrid,
-  ChartLegend,
-  ChartTooltip,
+  typedCartesianChart,
   Heading,
   Section,
-  XAxis,
-  YAxis,
 } from "@mittwald/flow-react-components";
 
-<Section>
-  <Heading>CPU</Heading>
-  <CartesianChart
-    data={[
-      {
-        Zeit: "0 Uhr",
-        Datenbanken: 40,
-        Projekte: 24,
-      },
-      {
-        Zeit: "1 Uhr",
-        Datenbanken: 30,
-        Projekte: 13,
-      },
-      {
-        Zeit: "2 Uhr",
-        Datenbanken: 20,
-        Projekte: 78,
-      },
-      {
-        Zeit: "3 Uhr",
-        Datenbanken: 27,
-        Projekte: 39,
-      },
-    ]}
-    height="200px"
-    syncId="syncedCharts"
-  >
-    <CartesianGrid />
-    <Area dataKey="Datenbanken" unit="%" />
-    <Area
-      dataKey="Projekte"
-      color="palatinate-blue"
-      unit="%"
-    />
-    <XAxis dataKey="Zeit" />
-    <YAxis domain={[0, 100]} unit=" %" />
-    <ChartTooltip />
-  </CartesianChart>
-  <Heading>RAM</Heading>
-  <CartesianChart
-    data={[
-      {
-        Zeit: "0 Uhr",
-        Datenbanken: 39,
-        Projekte: 27,
-      },
-      {
-        Zeit: "1 Uhr",
-        Datenbanken: 30,
-        Projekte: 13,
-      },
-      {
-        Zeit: "2 Uhr",
-        Datenbanken: 63,
-        Projekte: 10,
-      },
-      {
-        Zeit: "3 Uhr",
-        Datenbanken: 24,
-        Projekte: 40,
-      },
-    ]}
-    syncId="syncedCharts"
-  >
-    <CartesianGrid />
-    <Area dataKey="Datenbanken" unit="%" />
-    <Area
-      dataKey="Projekte"
-      color="palatinate-blue"
-      unit="%"
-    />
-    <XAxis dataKey="Zeit" />
-    <YAxis domain={[0, 100]} unit=" %" />
-    <ChartLegend />
-    <ChartTooltip />
-  </CartesianChart>
-</Section>;
+export default () => {
+  interface ChartData {
+    Zeit: string;
+    Datenbanken: number;
+    Projekte: number;
+  }
+
+  const CartesianChart = typedCartesianChart<ChartData>();
+
+  return (
+    <Section>
+      <Heading>CPU</Heading>
+      <CartesianChart.Chart
+        data={[
+          {
+            Zeit: "0 Uhr",
+            Datenbanken: 40,
+            Projekte: 24,
+          },
+          {
+            Zeit: "1 Uhr",
+            Datenbanken: 30,
+            Projekte: 13,
+          },
+          {
+            Zeit: "2 Uhr",
+            Datenbanken: 20,
+            Projekte: 78,
+          },
+          {
+            Zeit: "3 Uhr",
+            Datenbanken: 27,
+            Projekte: 39,
+          },
+        ]}
+        height="200px"
+        syncId="syncedCharts"
+      >
+        <CartesianChart.Area
+          dataKey="Datenbanken"
+          unit="%"
+        />
+        <CartesianChart.Area
+          dataKey="Projekte"
+          color="palatinate-blue"
+          unit="%"
+        />
+        <CartesianChart.XAxis dataKey="Zeit" />
+        <CartesianChart.YAxis domain={[0, 100]} unit=" %" />
+        <CartesianChart.Grid />
+        <CartesianChart.Tooltip />
+      </CartesianChart.Chart>
+      <Heading>RAM</Heading>
+      <CartesianChart.Chart
+        data={[
+          {
+            Zeit: "0 Uhr",
+            Datenbanken: 39,
+            Projekte: 27,
+          },
+          {
+            Zeit: "1 Uhr",
+            Datenbanken: 30,
+            Projekte: 13,
+          },
+          {
+            Zeit: "2 Uhr",
+            Datenbanken: 63,
+            Projekte: 10,
+          },
+          {
+            Zeit: "3 Uhr",
+            Datenbanken: 24,
+            Projekte: 40,
+          },
+        ]}
+        syncId="syncedCharts"
+      >
+        <CartesianChart.Area
+          dataKey="Datenbanken"
+          unit="%"
+        />
+        <CartesianChart.Area
+          dataKey="Projekte"
+          color="palatinate-blue"
+          unit="%"
+        />
+        <CartesianChart.XAxis dataKey="Zeit" />
+        <CartesianChart.YAxis domain={[0, 100]} unit=" %" />
+        <CartesianChart.Grid />
+        <CartesianChart.Legend />
+        <CartesianChart.Tooltip />
+      </CartesianChart.Chart>
+    </Section>
+  );
+};