+++ packages/components/src/components/CartesianChart/components/CartesianGrid/CartesianGrid.tsx
@@ -1,24 +1,10 @@
-import type { FC } from "react";
-import * as Recharts from "recharts";
-import { useDesignTokens } from "../../../../lib/theming";
+import { ChartGrid, type ChartGridProps } from "@/components/CartesianChart";
 
-export type CartesianGridProps = Pick<
-  Recharts.CartesianGridProps,
-  "className" | "vertical" | "horizontal" | "strokeDasharray"
->;
+/** @deprecated Use ChartGridProps */
+export type CartesianGridProps = ChartGridProps;
 
-/** @flr-generate all */
-export const CartesianGrid: FC<CartesianGridProps> = (props) => {
-  const { vertical = false, ...rest } = props;
-  const designTokens = useDesignTokens();
-  return (
-    <Recharts.CartesianGrid
-      vertical={vertical}
-      {...rest}
-      stroke={designTokens["cartesian-grid"].color.value}
-      strokeWidth={designTokens["cartesian-grid"]["stroke-width"].value}
-    />
-  );
-};
-
-export default CartesianGrid;
+/**
+ * @deprecated Use ChartGrid
+ * @flr-generate all
+ */
+export const CartesianGrid = ChartGrid;