+++ packages/components/src/components/CartesianChart/components/ChartTooltip/TooltipLegendItem.tsx
@@ -7,6 +7,7 @@ import type {
   ValueType,
 } from "recharts/types/component/DefaultTooltipContent";
 import { usePromise } from "@mittwald/react-use-promise";
+import type { ChartDataValue } from "@/components/CartesianChart/CartesianChart";
 
 /** @internal */
 export type TooltipPayloadItem = NonNullable<
@@ -14,8 +15,8 @@ export type TooltipPayloadItem = NonNullable<
 >[number];
 
 /** @internal */
-interface LegendItemLabelProps extends Pick<
-  WithTooltipFormatters,
+interface LegendItemLabelProps<TData = ChartDataValue> extends Pick<
+  WithTooltipFormatters<TData>,
   "formatter"
 > {
   item: TooltipPayloadItem;
@@ -36,7 +37,7 @@ export const TooltipLegendItem: FC<LegendItemLabelProps> = ({
         return `${dataKey} (${value}${unit ? ` ${unit}` : ""})`;
       }
 
-      return formatter(value, dataKey, index, unit);
+      return formatter(value, dataKey as never, index, unit);
     },
     [value, dataKey, index, unit, formatter] as const,
   );