+++ packages/components/src/components/CartesianChart/components/YAxis/YAxis.tsx
@@ -1,11 +1,19 @@
 import type { FC } from "react";
 import * as Recharts from "recharts";
+import type {
+  ChartDataValue,
+  DataKey,
+  DataKeyValue,
+} from "@/components/CartesianChart/CartesianChart";
 import { useDesignTokens } from "../../../../lib/theming";
 
-export type YAxisProps = Pick<
+export type YAxisProps<
+  TData = ChartDataValue,
+  TDataKey extends DataKey<TData> = DataKey<TData>,
+  TDataMatch = DataKeyValue<TData, TDataKey>,
+> = Pick<
   Recharts.YAxisProps,
   | "className"
-  | "dataKey"
   | "orientation"
   | "allowDecimals"
   | "interval"
@@ -15,8 +23,10 @@ export type YAxisProps = Pick<
   | "domain"
   | "hide"
   | "unit"
-  | "tickFormatter"
->;
+> & {
+  dataKey?: TDataKey;
+  tickFormatter?: (value: TDataMatch, index: number) => string;
+};
 
 /** @flr-generate all */
 export const YAxis: FC<YAxisProps> = (props) => {
@@ -38,5 +48,7 @@ export const YAxis: FC<YAxisProps> = (props) => {
     />
   );
 };
+
+export { TypedYAxis } from "./types";
 
 export default YAxis;