+++ packages/components/src/components/CartesianChart/index.ts
@@ -4,10 +4,14 @@ export {
   type CartesianChartProps,
   CartesianChart,
   type CartesianChartEmptyViewProps,
+  typedCartesianChart,
 } from "./CartesianChart";
+export * from "./view";
 export * from "./components/Area";
+export * from "./components/AreaDot";
 export * from "./components/Line";
 export * from "./components/ChartTooltip";
+export * from "./components/ChartGrid";
 export * from "./components/CartesianGrid";
 export * from "./components/ChartLegend";
 export * from "./components/YAxis";