+++ packages/ext-bridge/src/integrations/i18next/modules/remoteLanguageDetectorModule.ts
@@ -1,11 +1,15 @@
 import { getConfig } from "@/index-browser";
+import { isBrowser } from "@/lib/assertBrowserEnv";
 import { type LanguageDetectorAsyncModule } from "i18next";
 
-export const remoteLanguageDetectorModule: LanguageDetectorAsyncModule = {
-  type: "languageDetector",
-  async: true,
-  detect: async () => {
-    const config = await getConfig();
-    return config.language;
-  },
-};
+export const remoteLanguageDetectorModule =
+  (): LanguageDetectorAsyncModule => ({
+    type: "languageDetector",
+    async: true,
+    detect: async () => {
+      if (isBrowser) {
+        const config = await getConfig();
+        return config.language;
+      }
+    },
+  });